+++ components/bridge/rpc/impl/src/main/java/com/flipperdevices/bridge/rpc/impl/api/FlipperStorageApiImpl.kt
@@ -1,6 +1,7 @@
 package com.flipperdevices.bridge.rpc.impl.api
 
 import com.flipperdevices.bridge.api.model.wrapToRequest
+import com.flipperdevices.bridge.api.utils.Constants
 import com.flipperdevices.bridge.rpc.api.FlipperStorageApi
 import com.flipperdevices.bridge.rpc.api.model.NameWithHash
 import com.flipperdevices.bridge.rpc.impl.delegates.FlipperListingDelegate
@@ -8,6 +9,8 @@ import com.flipperdevices.bridge.rpc.impl.delegates.FlipperUploadDelegate
 import com.flipperdevices.bridge.rpc.impl.delegates.MkDirDelegate
 import com.flipperdevices.bridge.service.api.provider.FlipperServiceProvider
 import com.flipperdevices.core.di.AppGraph
+import com.flipperdevices.core.log.LogTagProvider
+import com.flipperdevices.core.log.info
 import com.flipperdevices.core.progress.ProgressListener
 import com.flipperdevices.protobuf.Flipper
 import com.flipperdevices.protobuf.main
@@ -25,7 +28,9 @@ class FlipperStorageApiImpl @Inject constructor(
     private val mkDirDelegate: MkDirDelegate,
     private val flipperUploadDelegate: FlipperUploadDelegate,
     private val listingDelegate: FlipperListingDelegate
-) : FlipperStorageApi {
+) : FlipperStorageApi, LogTagProvider {
+    override val TAG = "FlipperStorageApi"
+
     override suspend fun mkdirs(path: String) = withContext(Dispatchers.Default) {
         mkDirDelegate.mkdir(flipperServiceProvider.getServiceApi().requestApi, path)
     }
@@ -70,14 +75,15 @@ class FlipperStorageApiImpl @Inject constructor(
 
     override suspend fun listingDirectoryWithMd5(pathOnFlipper: String): List<NameWithHash> {
         val serviceApi = flipperServiceProvider.getServiceApi()
-        //val md5ListingSupported = serviceApi.flipperVersionApi.isSupported(Constants.API_SUPPORTED_MD5_LISTING)
-        val md5ListingSupported = true
+        val md5ListingSupported = serviceApi.flipperVersionApi.isSupported(Constants.API_SUPPORTED_MD5_LISTING)
         return if (md5ListingSupported) {
+            info { "Use new md5 request api" }
             listingDelegate.listingWithMd5(
                 requestApi = serviceApi.requestApi,
                 pathOnFlipper = pathOnFlipper
             )
         } else {
+            info { "Use deprecated md5 request api" }
             listingDelegate.listingWithMd5Deprecated(
                 requestApi = serviceApi.requestApi,
                 pathOnFlipper = pathOnFlipper