+++ components/bridge/rpc/impl/src/main/java/com/flipperdevices/bridge/rpc/impl/delegates/FlipperListingDelegate.kt
@@ -6,18 +6,31 @@ import com.flipperdevices.bridge.rpc.api.model.NameWithHash
 import com.flipperdevices.core.ktx.jre.pmap
 import com.flipperdevices.protobuf.Flipper
 import com.flipperdevices.protobuf.main
+import com.flipperdevices.protobuf.storage.Storage
 import com.flipperdevices.protobuf.storage.listRequest
 import com.flipperdevices.protobuf.storage.md5sumRequest
-import javax.inject.Inject
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.toList
+import java.io.File
+import javax.inject.Inject
+
+private const val SIZE_BYTES_LIMIT = 10 * 1024 * 1024 // 10MiB
 
 class FlipperListingDelegate @Inject constructor() {
     suspend fun listing(requestApi: FlipperRequestApi, pathOnFlipper: String): List<String> {
+        return listingInternal(requestApi, pathOnFlipper, includeMd5Flag = false).map { it.name }
+    }
+
+    private suspend fun listingInternal(
+        requestApi: FlipperRequestApi,
+        pathOnFlipper: String,
+        includeMd5Flag: Boolean
+    ): List<Storage.File> {
         return requestApi.request(
             main {
                 storageListRequest = listRequest {
                     path = pathOnFlipper
+                    includeMd5 = includeMd5Flag
                 }
             }.wrapToRequest()
         ).toList().mapNotNull { response ->
@@ -28,40 +41,45 @@ class FlipperListingDelegate @Inject constructor() {
             } else {
                 error("Can't find storage list response, $response")
             }
-        }.flatten().map { it.name }
+        }.flatten()
     }
 
     suspend fun listingWithMd5(
         requestApi: FlipperRequestApi,
         pathOnFlipper: String
     ): List<NameWithHash> {
-        return requestApi.request(
-            main {
-                storageListRequest = listRequest {
-                    path = pathOnFlipper
-                    includeMd5 = true
-                }
-            }.wrapToRequest()
-        ).toList().mapNotNull { response ->
-            if (response.commandStatus != Flipper.CommandStatus.OK) {
-                error("Listing request failed for $pathOnFlipper with $response")
-            } else if (response.hasStorageListResponse()) {
-                response.storageListResponse.fileList
-            } else {
-                error("Can't find storage list response, $response")
+        return listingInternal(
+            requestApi,
+            pathOnFlipper,
+            includeMd5Flag = true
+        ).mapNotNull {
+            if (it.md5Sum.isNullOrBlank()) {
+                return@mapNotNull null
             }
-        }.flatten().map { NameWithHash(it.name, it.md5Sum) }
+            NameWithHash(it.name, it.md5Sum, it.size, it.type)
+        }
     }
 
     suspend fun listingWithMd5Deprecated(
         requestApi: FlipperRequestApi,
         pathOnFlipper: String
     ): List<NameWithHash> {
-        return listing(requestApi, pathOnFlipper).pmap { fileForMd5 ->
+        return listingInternal(
+            requestApi,
+            pathOnFlipper,
+            includeMd5Flag = false
+        ).pmap { fileForMd5 ->
+            if (fileForMd5.type != Storage.File.FileType.FILE) {
+                return@pmap null
+            }
+            if (fileForMd5.size > SIZE_BYTES_LIMIT) {
+                return@pmap null
+            }
+
             val md5Response = requestApi.request(
                 main {
                     storageMd5SumRequest = md5sumRequest {
-                        path = fileForMd5
+                        path = File(pathOnFlipper, fileForMd5.name).path
                     }
                 }.wrapToRequest()
             ).first()
@@ -72,9 +90,11 @@ class FlipperListingDelegate @Inject constructor() {
                 error("Can't find md5 response in $md5Response for $fileForMd5")
             }
             NameWithHash(
-                name = fileForMd5,
-                md5 = md5Response.storageMd5SumResponse.md5Sum
+                name = fileForMd5.name,
+                md5 = md5Response.storageMd5SumResponse.md5Sum,
+                size = fileForMd5.size,
+                type = fileForMd5.type
             )
-        }
+        }.filterNotNull()
     }
 }